ojas_journal/requirements.txt
asgiref==3.3.4
Django==3.2
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
mysqlclient==2.0.3
Pillow==8.2.0
python-dotenv==0.17.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
